Split transient detection into multiple files and fix FFT processor tests
